Updated the academy cinemas page with JS and bootstrap
Updated the page so there is a modal form that will allow you to input your information to get movie deals. Added a onclick button that will make a popover effect that can be clicked away. Also added ID's to the movies so that when you click on the name it'll take you to the movie